+++ lib/common/mixins/base_mixin.dart
@@ -28,12 +28,20 @@ mixin BaseMixin<T extends StatefulWidget> on State<T> {
   void addAppErrLog(String content) => addLog(content, color: Colors.red);
 
   /// 记录发送类日志
-  void addSendLog(String content, {Object? attachment, String? tag}) =>
-      addLog(content, color: Colors.green, attachment: attachment, tag: tag);
+  void addSendLog(
+    String content, {
+    Color color = Colors.green,
+    Object? attachment,
+    String? tag,
+  }) => addLog(content, color: color, attachment: attachment, tag: tag);
 
   /// 记录接收类日志
-  void addReceiveLog(String content, {Object? attachment, String? tag}) =>
-      addLog(content, color: Colors.blue, attachment: attachment, tag: tag);
+  void addReceiveLog(
+    String content, {
+    Color color = Colors.blue,
+    Object? attachment,
+    String? tag,
+  }) => addLog(content, color: color, attachment: attachment, tag: tag);
 
   /// 获取 asset 文件路径，用于发送媒体消息
   Future<String> getAssetFilePath(String assetPath) async {